refactor(core): move buildShapePath + helpers from pptx into core
Extracts the OOXML preset-geometry path builder out of the pptx
renderer and into a standalone module under @silurus/ooxml-core, so
the docx renderer can share the same prstGeom coverage instead of
re-implementing each preset shape (currently docx only supports rect
and line via parser-side `preset_subpaths`).

Moved: buildShapePath, drawStar, drawPolygon, ooxmlArcTo. ~2050 lines
of switch-case and arc math, with no semantics changes — just relocation
plus `export` annotations and a JSDoc clarification on the entry point.

The pptx renderer now imports them from core. No behavior change is
expected; pptx VRT confirms 69/69 pass with `demo/sample-1` 100%.

The next step (deferred — schema change) is to teach the docx parser
to emit `prstGeom` shape names + adj values rather than collapsing
everything to a custGeom subpath, then call buildShapePath from the
docx renderer.
